+++ gehennabot/sources.py
@@ -1,5 +1,5 @@
 from collections import namedtuple
-from typing import List
+from typing import Dict, List
 import json
 
 import requests
@@ -30,34 +30,23 @@ def slots_from_vtescards(url: str) -> List[Slot]:
         + '/export?type=LACKEY'
     )
     if r.status_code != 200:
-        return []
+        print('Deck não encontrado.')
+        return {}
     texto = r.text
     lines = [line for line in texto.split('\n') if '\t' in line]
     slots = [Slot(*line.split('\t')) for line in lines]
     return slots
 
 
-def slots_from_vdb(url: str) -> List[Slot]:
+def slots_from_vdb(url: str) -> Dict:
     deckid = url.replace('https://vdb.im/decks/', '')
     url = URL_API_DECK + deckid
     r = requests.get(url)
     if r.status_code != 200:
-        return []
+        print('Deck não encontrado.')
+        return {}
     cards = r.json().get('cards')
-    # TODO pegar o nome das cartas nos arquivos json
-    data_cards = data_crypt  = data_library = {}
-    with open(config.CRYPT_FILE_PATH) as file:
-        data_crypt = json.load(file)
-    data_cards.update(data_crypt)
-    with open(config.LIBRARY_FILE_PATH) as file:
-        data_library = json.load(file)
-    data_cards.update(data_library)
-    slots = []
-    for key in cards.keys():
-        card = data_cards[key]
-        slot = Slot(cards[key], card['Name'])
-        slots.append(slot)
-    return slots
+    return cards 
 
 
 def slots_from_amaranth(url: str) -> list[Slot]: