+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 # Changelog
 
-
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -8,11 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-
 - Mostrar imagens das cartas
+
+## [0.8.0]
+
+- Cadastrar Movimento a partir de um deck
 - Importar decks de outros sites
-- Enviar itens das movimentações
 - Usar API do Gehenna-API
+- Enviar itens das movimentações
 
 ## [0.7.0]
 
@@ -20,22 +22,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Comandos CLI para auxiliar em algumas tarefas
 
 ## [0.6.0]
+
 - Adoção do Poetry
 
 ## [0.5.1]
+
 - Corrigir o método service.composicao_por_id que não leva em conta o parâmetro usuário
 
 ## [0.5.0]
+
 - Sugestão de nomes para decks cadastrados via URL
 
 ## [0.4.0]
+
 - Cadastro de deck por URL (https://vtesdecks.com/) no formato Lackey
 
 ## [0.3.0]
+
 ### Adicionado
+
 - Listar os decks construídos que possuem cartas de um deck cadastrado
   - comando: /onde [deck_id]
 
 ## [0.2.0]
+
 ### Adicionado
+
 - Lista de cartas faltantes para completar um deck